+++ src/components/signup/Password.jsx
@@ -3,26 +3,7 @@ import { useState } from 'react';
 
 import Required from '../_common/Required';
 import Alert from './Alert';
-
-const isValidPassword = (password) => {
-  if (password.length < 6 || password.length > 20) {
-    return false;
-  }
-
-  const hasUpperCase = /[A-Z]/.test(password);
-  const hasLowerCase = /[a-z]/.test(password);
-  const hasNumber = /\d/.test(password);
-  const hasSpecialChar = /[!@#$%^&*()-_]/.test(password);
-
-  const conditionCount = [
-    hasUpperCase,
-    hasLowerCase,
-    hasNumber,
-    hasSpecialChar,
-  ].filter(Boolean).length;
-
-  return conditionCount >= 2;
-};
+import { isValidPassword } from '../../util/isValidPassword';
 
 const Password = () => {
   const [pw, setPw] = useState('');
