+++ src/util/isValidPassword.js
@@ -1,0 +1,19 @@
+export const isValidPassword = (password) => {
+  if (password.length < 6 || password.length > 20) {
+    return false;
+  }
+
+  const hasUpperCase = /[A-Z]/.test(password);
+  const hasLowerCase = /[a-z]/.test(password);
+  const hasNumber = /\d/.test(password);
+  const hasSpecialChar = /[!@#$%^&*()-_]/.test(password);
+
+  const conditionCount = [
+    hasUpperCase,
+    hasLowerCase,
+    hasNumber,
+    hasSpecialChar,
+  ].filter(Boolean).length;
+
+  return conditionCount >= 2;
+};
